imgengine-saas/backend/api-service/pyproject.toml
[project]
name = "imgengine"
version = "0.1.0"
description = "ImgEngine SAAS with fastapi and C "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.135.2",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
    "requests>=2.33.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "prometheus-client==0.20.0",
    # 🔥 ADD THESE
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-requests",
    "httpx>=0.28.1",
]
